requirements.txt
scikit_image==0.14.1
opencv_python==3.4.4.19
tqdm==4.28.1
torch==1.4.0
scipy==1.2.0
numpy==1.15.4
torchvision==0.2.1
Pillow==8.1.1
tensorboardX==1.5
editdistance==0.5.3
